Double to BigDecimal

+++ src/main/java/pl/kluczewski/currency_converter/controller/CurrencyController.java
@@ -6,6 +6,7 @@ import pl.kluczewski.currency_converter.model.AllCurrencyDto;
 import pl.kluczewski.currency_converter.model.CurrencyDto;
 import pl.kluczewski.currency_converter.service.CurrencyService;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @CrossOrigin
@@ -22,22 +23,22 @@ public class CurrencyController {
     }
 
     @GetMapping("/from/{currency}/{quantity}")
-    public CurrencyDto getValueFromPln(@PathVariable String currency, @PathVariable double quantity) {
+    public CurrencyDto getValueFromPln(@PathVariable String currency, @PathVariable BigDecimal quantity) {
         return currencyService.getValueFromPln(currency, quantity);
     }
 
     @GetMapping("/from/{currency}/{quantity}/{date}")
-    public CurrencyDto getValueFromPln(@PathVariable String currency, @PathVariable double quantity, @PathVariable String date) {
+    public CurrencyDto getValueFromPln(@PathVariable String currency, @PathVariable BigDecimal quantity, @PathVariable String date) {
         return currencyService.getValueFromPln(currency, quantity, date);
     }
 
     @GetMapping("/to/{currency}/{quantity}")
-    public CurrencyDto getValueToPln(@PathVariable String currency, @PathVariable double quantity) {
+    public CurrencyDto getValueToPln(@PathVariable String currency, @PathVariable BigDecimal quantity) {
         return currencyService.getValueToPln(currency, quantity);
     }
 
     @GetMapping("/to/{currency}/{quantity}/{date}")
-    public CurrencyDto getValueToPln(@PathVariable String currency, @PathVariable double quantity, @PathVariable String date) {
+    public CurrencyDto getValueToPln(@PathVariable String currency, @PathVariable BigDecimal quantity, @PathVariable String date) {
         return currencyService.getValueToPln(currency, quantity, date);
     }
 }
