+++ src/main/java/pl/kluczewski/currency_converter/service/CurrencyService.java
@@ -6,6 +6,7 @@ import pl.kluczewski.currency_converter.model.AllCurrencyDto;
 import pl.kluczewski.currency_converter.model.CurrencyDto;
 import pl.kluczewski.currency_converter.webclient.CurrencyClient;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -19,19 +20,19 @@ public class CurrencyService {
         return currencyClient.getValueForAllCurrency();
     }
 
-    public CurrencyDto getValueFromPln(String currency, double quantity) {
+    public CurrencyDto getValueFromPln(String currency, BigDecimal quantity) {
         return currencyClient.getValueFromPln(currency, quantity);
     }
 
-    public CurrencyDto getValueFromPln(String currency, double quantity, String date) {
+    public CurrencyDto getValueFromPln(String currency, BigDecimal quantity, String date) {
         return currencyClient.getValueFromPln(currency, quantity, date);
     }
 
-    public CurrencyDto getValueToPln(String currency, double quantity) {
+    public CurrencyDto getValueToPln(String currency, BigDecimal quantity) {
         return currencyClient.getValueToPln(currency, quantity);
     }
 
-    public CurrencyDto getValueToPln(String currency, double quantity, String date) {
+    public CurrencyDto getValueToPln(String currency, BigDecimal quantity, String date) {
         return currencyClient.getValueToPln(currency, quantity, date);
     }
 }
