+++ src/main/java/pl/kluczewski/currency_converter/webclient/dto/RatesDto.java
@@ -2,10 +2,12 @@ package pl.kluczewski.currency_converter.webclient.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 
 @Getter
 public class RatesDto {
     private String no;
     private String effectiveDate;
-    private double mid;
+    private BigDecimal mid;
 }
